just lots of cleanup and docs
