+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: zcUOJV55Tjfuf82xwdXBlaztt3H1RawBf
