+++ .travis.yml
@@ -8,3 +8,5 @@ notifications:
     on_success: change
     on_failure: always
     on_start: false
+script: "npm run-script ci"
+after_script: "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
