+++ lib/field.js
@@ -1,23 +1,9 @@
-var es = require('event-stream');
 var _ = require('lodash');
-var raf = require('raf');
-
-var debug = require('debug')('mongodb-schema');
-
-var Collection = require('./collection');
 var State = require('./state');
-var Type = require('./type');
 var TypeCollection = require('./type-collection');
 var ValueCollection = require('./value-collection');
-
-/**
- * Container for a list of Fields.
- */
-var FieldCollection = Collection.extend({
-  model: function(attrs, options) {
-    return new attrs.klass(attrs, options);
-  }
-});
+var FieldCollection = require('./field-collection');
+var debug = require('debug')('mongodb-schema:field');
 
 /**
  * Like a property.
@@ -259,21 +245,15 @@ var Field = State.extend({
     return res;
   },
 });
+module.exports = Field;
 
 /**
  * A basic field has no descendant fields, such as `String`, `ObjectID`,
  * `Boolean`, or `Date`.
- * @class
- * @implements Field
  */
-var BasicField = Field.extend({});
+module.exports.Basic = Field.extend({});
 
-/**
- * An embedded array of Fields.
- * @class
- * @implements Field
- */
-var EmbeddedArrayField = Field.extend({
+module.exports.EmbeddedArray = Field.extend({
   props: {
     type: {
       type: 'string',
@@ -296,11 +276,7 @@ var EmbeddedArrayField = Field.extend({
   }
 });
 
-/**
- * @class
- * @implements Field
- */
-var EmbeddedDocumentField = Field.extend({
+module.exports.EmbeddedDocument = Field.extend({
   props: {
     type: {
       type: 'string',
@@ -308,129 +284,3 @@ var EmbeddedDocumentField = Field.extend({
     }
   }
 });
-
-function onFieldSampled(schema, _id, value) {
-  var type_id = Type.getNameFromValue(value);
-  if (type_id === 'Array') {
-    onEmbeddedArray(schema, _id, type_id, value);
-  } else if (type_id === 'Object') {
-    onEmbeddedDocument(schema, _id, type_id, value);
-  } else {
-    onBasicField(schema, _id, type_id, value);
-  }
-}
-
-function onBasicField(schema, _id, type_id, value) {
-  var field = schema.fields.get(_id);
-  if (!field) {
-    field = schema.fields.add({
-      _id: _id,
-      klass: BasicField,
-      parent: schema
-    });
-  }
-  field.count += 1;
-
-  var type = field.types.get(type_id);
-  if (!type) {
-    type = field.types.add({
-      _id: type_id,
-    });
-  }
-  type.count += 1;
-
-  type.values.add({
-    _id: value
-  });
-}
-
-function onEmbeddedArray(schema, _id, type_id, value) {
-  var field = schema.fields.get(_id);
-
-  if (!field) {
-    field = schema.fields.add({
-      _id: _id,
-      klass: EmbeddedArrayField,
-      parent: schema
-    });
-  }
-
-  field.count += 1;
-  field.lengths.push(value.length);
-  field.trigger('change:lengths');
-  _.each(value, function(d) {
-    var type_id = Type.getNameFromValue(d);
-    if (type_id === 'Object') {
-      _.each(d, function(val, key) {
-        onBasicField(field, key, Type.getNameFromValue(val), val);
-      });
-    } else {
-      onBasicField(field, '__basic__', type_id, d);
-    }
-  });
-}
-
-function onEmbeddedDocument(schema, _id, type_id, value) {
-  var field = schema.fields.get(_id);
-
-  if (!field) {
-    field = schema.fields.add({
-      _id: _id,
-      klass: EmbeddedDocumentField,
-      parent: schema
-    });
-  }
-  field.count += 1;
-  _.each(value, function(val, key) {
-    onFieldSampled(field, key, val);
-  });
-}
-
-/**
- * The top level schema state.
- * @class
- */
-var Schema = State.extend({
-  idAttribute: 'ns',
-  props: {
-    ns: {
-      type: 'string'
-    },
-    count: {
-      type: 'number',
-      default: 0
-    }
-  },
-  collections: {
-    fields: FieldCollection
-  },
-  parse: function(doc, done) {
-    var schema = this;
-    schema.count += 1;
-    _.each(doc, function(val, key) {
-      onFieldSampled(schema, key, val);
-    });
-    schema.fields.map(function(field) {
-      field.commit();
-    });
-    if (_.isFunction(done)) {
-      done();
-    }
-  },
-  stream: function() {
-    var schema = this;
-    return es.map(function(doc, done) {
-      raf(function() {
-        schema.parse(doc, function(err) {
-          done(err, doc);
-        });
-      });
-    });
-  }
-});
-
-module.exports = Schema;
-module.exports.FieldCollection = FieldCollection;
-module.exports.BasicField = BasicField;
-module.exports.EmbeddedArrayField = EmbeddedArrayField;
-module.exports.EmbeddedDocumentField = EmbeddedDocumentField;
