+++ lib/index.js
@@ -1,22 +1,30 @@
 var Schema = require('./schema');
 var es = require('event-stream');
-var assert = require('assert');
 
+/**
+ * Convenience shortcut for parsing schemas.
+ * @param {String} ns The namespace of the collection being parsed.
+ * @param {Cursor|Array} docs An array of documents or a Cursor returned by `.find()`
+ * @param {Function} fn Callback which will be passed `(err, schema)`
+ * @returns {Schema}
+ */
 module.exports = function(ns, docs, fn) {
-  assert(Array.isArray(docs), 'docs must be an array');
   var schema = new Schema({
     ns: ns
   });
+  var src;
 
-  es.readArray(docs).pipe(schema.stream()).on('end', fn);
+  if(docs.stream){
+    src = docs.stream();
+  }
+  else{
+    src = es.readArray(docs);
+  }
+
+  src.pipe(schema.stream()).on('end', function(){
+    fn.call(null, null, schema);
+  });
   return schema;
 };
 
-module.exports.extend = Schema.extend.bind(Schema);
 module.exports.Schema = Schema;
-module.exports.getType = require('./type').getNameFromValue;
-module.exports.FieldCollection = Schema.FieldCollection;
-module.exports.BasicField = Schema.BasicField;
-module.exports.EmbeddedArrayField = Schema.EmbeddedArrayField;
-module.exports.EmbeddedDocumentField = Schema.EmbeddedDocumentField;
-module.exports.TypeCollection = require('./type-collection');
