+++ lib/type-collection.js
@@ -1,14 +1,11 @@
 var Collection = require('./collection');
 var type = require('./type');
+var assert = require('assert');
 
 module.exports = Collection.extend({
   model: function(attrs, options) {
     var Klass = type[attrs._id];
-
-    if (!Klass) {
-      throw new TypeError('No value type for ' + attrs._id);
-    }
-
+    assert(Klass, 'No value type for ' + attrs._id);
     return new Klass(attrs, options);
   },
   /**
