+++ lib/value.js
@@ -12,8 +12,5 @@ module.exports = State.extend({
   initialize: function(attrs) {
     this.value = attrs._id;
     this._id = this.cid + '-' + attrs._id;
-  },
-  valueOf: function() {
-    return this.value;
   }
 });
